среднее доработано

+++ src/Main.java
@@ -12,7 +12,7 @@ public class Main {
     //первое задание
     public static void main(String[] args) {
         int[] arr = generateRandomArray();
-        int arrOne = 0;
+        double arrOne = 0;
         for (int i = 0; i < arr.length; i++) {
             arrOne += arr[i];
         }
@@ -31,8 +31,8 @@ public class Main {
         }
         System.out.println("Минимальная сумма трат за день составила " + arrTwo + " рублей. Максимальная сумма трат за день составила " + arrThree + " рублей");
         //Третье задание
-        float division = 0;
-        division = arrOne / arr.length;
+        double division = 0;
+        division = (double)arrOne / arr.length;
         System.out.println("Средняя сумма затрат за месяц составила " + division + " рублей");
         //Четвёртое задание
         char[] reverseFullName = { 'n', 'a', 'v', 'I', ' ', 'v', 'o', 'n', 'a', 'v', 'I'};
